requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
pydantic==2.10.3
email-validator==2.2.0
python-multipart==0.0.18
python-dotenv==1.0.1
